writing-pratice/requirements.txt
boto3>=1.34.0
youtube-transcript-api>=0.6.2
chromadb>=0.4.22
groq
streamlit
requests
python-dotenv